bigint: add custom gates for dot product of len 3 vector with constant len 3 vector
this reduced pairing cell count from 3.1M to 2.66M (for `21` lookup bits, `88`
limb bits) but seems to have made proving time slightly worse.

deleted the custom gates that used higher Rotation(..) since we knew it
was slower

it seems like vertical custom gates are just not the way to go?